Move persistence modules to /stores
